+++ solve/src/commonMain/kotlin/it/unibo/tuprolog/solve/SolverBuilder.kt
@@ -32,7 +32,7 @@ interface SolverBuilder {
     fun runtime(runtime: Runtime): SolverBuilder
 
     @JsName("builtins")
-    var builtins: Library
+    var builtins: Library?
 
     @JsName("setBuiltins")
     fun builtins(builtins: Library): SolverBuilder
